Backend/requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
dnspython==2.6.1
mongoengine==0.28.2
PyJWT==2.8.0
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
